stream-dist/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL support
aiosqlite==0.19.0       # Async SQLite support

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# WebSocket Support
python-socketio==5.10.0
python-engineio==4.8.0

# Process Management
psutil==5.9.6

# HTTP Client
httpx==0.25.1
aiohttp==3.9.1

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Date/Time
python-dateutil==2.8.2

# Logging
structlog==23.2.0
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1  # For testing async endpoints

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Development
watchfiles==0.21.0  # For auto-reload in development
